page.jsxでSuspenseを使用してMyPageをラップ
